requirements.txt
'''streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
tempfile-dir>=1.0.0'''
streamlit==1.28.1
google-generativeai==0.3.2
pandas==1.5.3
plotly==5.17.0